Adaptación CRM - Cotizador Programación CIFA

+++ api.imnc/imnc/cotizaciones/update/index.php
@@ -97,6 +97,22 @@ if($ID_TIPO_SERVICIO == 16){
 	}
 }
 
+//Solo para CIFA
+$MODALIDAD = "";
+$ID_CURSO = "";
+$CANT_PARTICIPANTES = 0;
+if($ID_SERVICIO == 3){
+	$MODALIDAD = $objeto->MODALIDAD;
+	valida_parametro_and_die($MODALIDAD,"Falta MODALIDAD");
+	$ID_CURSO = $objeto->ID_CURSO;
+	valida_parametro_and_die($ID_CURSO,"Falta ID CURSO");
+	//Para los cursos in situ validar que tenga cantidad de participantes
+	if($MODALIDAD == "insitu"){
+		$CANT_PARTICIPANTES = $objeto->CANT_PARTICIPANTES;
+		valida_parametro_and_die($CANT_PARTICIPANTES,"Falta CANT_PARTICIPANTES");
+	}
+}
+
 $ID_USUARIO_MODIFICACION = $objeto->ID_USUARIO;
 valida_parametro_and_die($ID_USUARIO_MODIFICACION,"Falta ID de USUARIO");
 
@@ -125,21 +141,54 @@ $id = $database->update("COTIZACIONES", [
 ], ["ID"=>$ID]); 
 
 valida_error_medoo_and_die(); 
-//Borrar todas las normas
-$id = $database->delete("COTIZACION_NORMAS", 
-	[
-		"AND" => [
-			"ID_COTIZACION" => $ID
-		]		
-	]);
-//iNSERTAR LAS NORMAS
-for ($i=0; $i < count($NORMAS); $i++) {
-	$id_norma = $NORMAS[$i]->ID_NORMA;
-	$id_cotizacion_normas = $database->insert("COTIZACION_NORMAS", [
-		"ID_COTIZACION" => $ID,
-		"ID_NORMA" => $id_norma
+
+if($ID_SERVICIO != 3){
+	//Borrar todas las normas
+	$id = $database->delete("COTIZACION_NORMAS", 
+		[
+			"AND" => [
+				"ID_COTIZACION" => $ID
+			]		
+		]);
+	//iNSERTAR LAS NORMAS
+	for ($i=0; $i < count($NORMAS); $i++) {
+		$id_norma = $NORMAS[$i]->ID_NORMA;
+		$id_cotizacion_normas = $database->insert("COTIZACION_NORMAS", [
+			"ID_COTIZACION" => $ID,
+			"ID_NORMA" => $id_norma
+		]);
+		valida_error_medoo_and_die();
+	}
+} else { //CIFA
+	//para CIFA insertar los detalles MODALIDAD y ID_CURSO
+	//Si MODALIDAD = programado ID_CURSO => columna ID_CURSO en CURSOS_PROGRAMADOS
+	//Si MODALIDAD = insitu ID_CURSO => columna ID_CURSO en CURSOS
+	$id_cotizacion_detalles = $database->update("COTIZACION_DETALLES", [
+		"VALOR"	=>	$MODALIDAD
+	],["AND" => [
+			"ID_COTIZACION" => $ID,
+			"DETALLE" => "MODALIDAD"
+		]
 	]);
 	valida_error_medoo_and_die();
+	$id_cotizacion_detalles = $database->update("COTIZACION_DETALLES", [
+		"VALOR"	=>	$ID_CURSO
+	],["AND"=>[
+			"ID_COTIZACION" => $ID,
+			"DETALLE" => "ID_CURSO"
+		]
+	]);
+	valida_error_medoo_and_die();
+	if($MODALIDAD == "insitu"){
+		$id_cotizacion_detalles = $database->update("COTIZACION_DETALLES", [
+			"VALOR"	=>	$CANT_PARTICIPANTES
+		],["AND"=>[
+				"ID_COTIZACION" => $ID,
+				"DETALLE" => "CANT_PARTICIPANTES"
+			]
+		]);
+		valida_error_medoo_and_die();
+	}
 }
 
 //Si la cotizacion tiene algun detalle que deba ser guardado en la tabla cotizacion detalles.
